Stimulation
A new function in the neural networks allows NSS to calculate the "stimulation" of an organism. The stimulation value shows how much stimulation the neural network is receiving.
If too much of the same inputs are being used, the stimulation will decrease - allowing you to tell how "bored" that creature is.